+++ database/src/main/java/com/mint/db/config/ConfigParser.java
@@ -1,8 +1,10 @@
 package com.mint.db.config;
 
+import org.yaml.snakeyaml.DumperOptions;
 import org.yaml.snakeyaml.LoaderOptions;
 import org.yaml.snakeyaml.Yaml;
 import org.yaml.snakeyaml.constructor.Constructor;
+import org.yaml.snakeyaml.representer.Representer;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -12,10 +14,17 @@ import java.io.InputStream;
 public final class ConfigParser {
     private static final String NODE_CONFIG_PATH_DEFAULT = "node-config.yaml";
     private static final String NODE_CONFIG_ENV_PROPERTY = "mint.config.location";
-    private static final Yaml YAML = new Yaml(new Constructor(NodeConfig.class, new LoaderOptions()));
+    private static final Yaml YAML;
+
+    static {
+        Representer representer = new Representer(new DumperOptions());
+        representer.getPropertyUtils().setSkipMissingProperties(true);
+        LoaderOptions loaderOptions = new LoaderOptions();
+        Constructor constructor = new Constructor(NodeConfig.class, loaderOptions);
+        YAML = new Yaml(constructor, representer);
+    }
 
     private ConfigParser() {
-
     }
 
     public static NodeConfig parseConfig() throws FileNotFoundException {
