+++ database/src/main/java/com/mint/db/dao/Dao.java
@@ -1,4 +1,4 @@
-package com.mint.db;
+package com.mint.db.dao;
 
 import java.util.Iterator;
