+++ database/src/main/java/com/mint/db/replication/model/LogEntry.java
@@ -1,6 +1,6 @@
 package com.mint.db.replication.model;
 
-import com.mint.db.Entry;
+import com.mint.db.dao.Entry;
 import com.mint.db.replication.model.impl.OperationType;
 
 public interface LogEntry<D> {
